+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Unreleased]
 ### Addded
 - Color palette chooser.
+- Multivariate histograms.
 
 ## [v0.3.1]
 ### Fixed