+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### [Unreleased]
+### [v0.3.2]
 ### Fixed
 - Allow axis label gap fix for numberical axis type.
 - Bug caused when there were spaces or other special characters in tab headers.